make release moved to DevTools folder
